requirements.txt
alabaster==0.7.12
astroid==2.2.5
Babel==2.6.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
codecov==2.0.15
colorclass==2.2.0
coverage==4.5.3
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-cors-headers==2.5.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.9.2
docopt==0.6.2
docutils==0.14
httpie==1.0.2
idna==2.8
imagesize==1.1.0
isort==4.3.17
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.2
oauthlib==3.0.1
olefile==0.46
packaging==19.0
Pillow==6.0.0
pip-upgrader==1.4.6
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.0
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
terminaltables==3.1.0
typed-ast==1.3.4
urllib3==1.24.2
wrapt==1.11.1
